split InteractionForm functionality from Modal
