+++ src/components/Forms/InteractionForm/index.ts
@@ -1,0 +1,2 @@
+export * from './InteractionForm';
+export * from './InteractionFormContent';
