+++ src/components/Forms/InteractionForm/styles.module.scss
@@ -1,4 +1,4 @@
-.interaction-modal-form {
+.interaction-form {
   /*Placeholder*/
   &-fields {
     /*Placeholder*/
