+++ src/components/Forms/types.ts
@@ -1,3 +1,3 @@
 export * from './DeployUpgradeFileForm/types';
-export * from './InteractionModalForm/types';
 export * from './EndpointForm/types';
+export * from './InteractionForm/types';
