+++ src/hooks/index.ts
@@ -3,6 +3,7 @@ export * from './provider';
 export * from './useGetAccountTokens';
 export * from './useGetTransaction';
 export * from './useGetTransactionCost';
+export * from './useGetTransactionCostDetails';
 export * from './useQueryContract';
 export * from './useTrackTransaction';
 export * from './useUpdateDeployedContractDetails';
